:art: moves Routes components to src/Routes
Moves the custom routing components to the 'src/' folder as they are high-level. This also removes the serviceWorker.js as we are not using it
